rag-chatbot/requirements.txt
# Core
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# Vector Store
chromadb>=0.5.0

# PDF Processing
pypdf>=4.0.0

# UI
streamlit>=1.38.0

# Environment
python-dotenv>=1.0.0
